SoilQuality-EHO/requirements.txt
fastapi
uvicorn
tensorflow==2.13.0
numpy
joblib
pillow
python-multipart
scikit-learn